# Improve the performance of ltrim/rtrim/btrim