@Omega359 2024-04-09
Nice! It would be a nice addition if the benchmark was expanded to cover btrim and rtrim as well