@JasonLi-cn 2024-04-10
> Nice! It would be a nice addition if the benchmark was expanded to cover btrim and rtrim as well