@JasonLi-cn 2024-04-10
> Nice! It would be a nice addition if the benchmark was expanded to cover btrim and rtrim as well

Thank you for your @Omega359 suggestion. I still need to ask @alamb whether it is necessary to add benchmarks for btrim/rtrim.