@alamb 2024-04-10
> > Nice! It would be a nice addition if the benchmark was expanded to cover btrim and rtrim as well
> 
> Thank you @Omega359 for your suggestion. I still need to ask @alamb whether it is necessary to add benchmarks for btrim/rtrim.

it is not necessary, though it would be nice as @Omega359  said. We can also do it as a follow on PR. Thanks again @JasonLi-cn